Create base ProfileStore
