+++ packages/yoroi-extension/app/containers/settings/categories/GeneralSettingsPage.stories.js
@@ -10,7 +10,7 @@ import GeneralSettingsPage from './GeneralSettingsPage';
 import { withScreenshot } from 'storycap';
 import { globalKnobs, } from '../../../../stories/helpers/StoryWrapper';
 import { walletLookup } from '../../../../stories/helpers/WalletCache';
-import { getVarsForTheme } from '../../../stores/toplevel/ProfileStore';
+import { getVarsForTheme } from '../../../stores/base/BaseProfileStore';
 import { wrapSettings } from '../../../Routes';
 import { mockSettingsProps } from '../Settings.mock';
 import { ROUTES } from '../../../routes-config';
