+++ packages/yoroi-extension/app/ergo-connector/actions/index.js
@@ -1,18 +1,18 @@
 // @flow
-import ProfileActions from '../../actions/profile-actions';
+import BaseProfileActions from '../../actions/base/base-profile-actions';
 import DialogsActions from '../../actions/dialogs-actions';
 import NotificationsActions from '../../actions/notifications-actions';
 import ConnectorActions from './connector-actions';
 
 export type ActionsMap = {|
-  profile: ProfileActions,
+  profile: BaseProfileActions,
   dialogs: DialogsActions,
   notifications: NotificationsActions,
   connector: ConnectorActions,
 |};
 
 const actionsMap: ActionsMap = Object.freeze({
-  profile: new ProfileActions(),
+  profile: new BaseProfileActions(),
   connector: new ConnectorActions(),
   dialogs: new DialogsActions(),
   notifications: new NotificationsActions(),
